Add minor comment for vTaskPriorityGet()
In ch04, listing 4.22, there was minor comment
missing about getting a task handle from the
xTaskCreateStatic() function.
As in vTaskPrioritySet() prototype description
there is such comment, it could be misleading that
there is any difference for vTaskPriorityGet().
So it is better if the comment is added also in
its prototype description.