Improve SW timer changing period example
In the example in Listing 6.13, there is a flag,
which should indicate that an error has been
detected by CheckTasksAreRunningWithoutError().
This flag was always set, even if the result from
the function call is not pdFAIL. For correct
usage of the flag it should be placed inside the
if condition.